Function: Moving all outward-facing `Function` types to a submodule
