+++ devito/__init__.py
@@ -8,7 +8,8 @@ from devito.backends import init_backend
 from devito.finite_difference import *  # noqa
 from devito.dimension import *  # noqa
 from devito.grid import *  # noqa
-from devito.interfaces import Forward, Backward, _SymbolCache  # noqa
+from devito.function import Forward, Backward  # noqa
+from devito.interfaces import _SymbolCache  # noqa
 from devito.logger import error, warning, info  # noqa
 from devito.parameters import (configuration, init_configuration,  # noqa
                                env_vars_mapper)
