+++ devito/foreign/__init__.py
@@ -6,6 +6,5 @@ Further, with the ``foreign`` backed, Devito doesn't allocate any data.
 """
 
 # The following used by backends.backendSelector
-from devito.interfaces import Constant, Function, TimeFunction  # noqa
-from devito.sparsefunction import SparseFunction  # noqa
+from devito.function import Constant, Function, TimeFunction, SparseFunction  # noqa
 from devito.foreign.operator import Operator  # noqa
